apiTest 수정 및 test coverage 가능하도록 수정
