--- AI/src/hand_detection/google_hand.py
@@ -1,16 +1,0 @@
-import cv2
-import mediapipe as mp
-mp_drawing = mp.solutions.drawing_utils
-mp_drawing_styles = mp.solutions.drawing_styles
-mp_hands = mp.solutions.hands
-
-def google_hands(file):
-    # For static images:
-    with mp_hands.Hands(
-        static_image_mode=True,
-        max_num_hands=4,
-        min_detection_confidence=0.3) as hands:
-        image = cv2.flip(file, 1)
-        # Convert the BGR image to RGB before processing.
-        results = hands.process(cv2.cvtColor(image, cv2.COLOR_BGR2RGB))
-        return len(results.multi_handedness)
