+++ AI/test/all_test.sh
@@ -3,17 +3,20 @@
 # api Test
 echo API TEST
 echo hand detection TEST
-python AI/test/apiTest/hand_detection_apiTest.py
+coverage run AI/test/apiTest/apiTest.py
 echo "\n"
-echo identification TEST
-python AI/test/apiTest/identification_apiTest.py
-
+coverage report -m AI/src/*.py
 
 # module Test
 echo "\n\n\n"
 echo MODULE TEST
 echo hand detection TEST
-python AI/test/moduleTest/hand_detection_moduleTest.py
+coverage run AI/test/moduleTest/hand_detection_moduleTest.py
+echo "\n"
+coverage report -m AI/src/hand_detection/*.py
+
 echo "\n"
 echo identification TEST
-python AI/test/moduleTest/identification_moduleTest.py
+coverage run AI/test/moduleTest/identification_moduleTest.py
+echo "\n"
+coverage report -m AI/src/student_identification/*.py
